+++ src/api/caidapp/urls.py
@@ -71,6 +71,9 @@ urlpatterns = [
     path("albums/", views.albums, name="albums"),
     path("album/<str:album_hash>", views.media_files_update, name="album"),
     path("taxon/<int:taxon_id>", views.media_files_update, name="taxon"),
+    path("representative_mediafiles/",
+         lambda request: views.media_files_update(request, identity_is_representative=True),
+         name="representative_mediafiles"),
     path(
         "individual_identity_mediafiles/<int:individual_identity_id>",
         views.media_files_update,