+++ changelog.md
@@ -9,6 +9,7 @@
 * [fixed] Re-run of taxon classification works again for workgroup users
 * [added] Paired points from LOFTR are now shown in the detail view
 * [changed] gpu:1 used for SAM
+* [added] List all representative files for identification
 
 ## 03-2024